Update DoomsDayMain.py
switched from Flask to Twisted, can serve a static directory with Twisted, and I know how to do such a thing with Twisted. Flask is too much of a learning curve at the moment. Working on a running random seed every X seconds to make it appear as though there are new people joining the website(give the appearance that its a highly trafficked site, or that the site is active, and incentivize the attacker to try to "steal" the database even more.